consistent use of loss registry system
