+++ utils/loss/losses.py
@@ -1,0 +1,205 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+import torch.distributed as dist
+from typing import Any, Dict, Optional
+
+from utils.registry import LossRegistry
+
+
+def get_loss_fn(
+    loss_name: str, 
+    kwargs: Optional[Dict[str, Any]] = None
+) -> nn.Module:
+    """
+    Returns the appropriate loss function by name.
+    """
+    if kwargs is None:
+        kwargs = {}
+    return LossRegistry.get(loss_name)(**kwargs)
+
+@LossRegistry.register('contrastive')
+class ContrastiveLoss(nn.Module):
+    """
+    CLIP-style bidirectional contrastive loss for video and text embeddings.
+    """
+    def __init__(self):
+        """
+        Initialize contrastive loss.
+        """
+        super().__init__()
+    
+    def forward(
+        self,
+        video_features: torch.Tensor, 
+        text_features: torch.Tensor,
+        log_temp: torch.Tensor = torch.log(torch.tensor(0.1))
+    ) -> torch.Tensor:
+        """
+        Compute contrastive loss between video and text features.
+        
+        Args:
+            video_features (torch.Tensor): [B, D] video embeddings
+            text_features (torch.Tensor): [B, D] text embeddings
+            log_temp (torch.Tensor, optional): float torch tensor value
+            
+        Returns:
+            torch.Tensor: Scalar loss value
+        """        
+        # Normalize embeddings.
+        video_features = F.normalize(video_features, dim=1)
+        text_features = F.normalize(text_features, dim=1)
+        
+        # Compute similarity matrix of shape [B, B].
+        similarity_matrix = torch.matmul(video_features, text_features.t())
+        
+        # Apply temperature scaling.
+        temp = torch.exp(log_temp)
+        logits = similarity_matrix / temp
+
+        # Create targets: assume the i-th video corresponds to the i-th text.
+        targets = torch.arange(logits.size(0), device=logits.device)
+
+        # Compute cross-entropy loss in both directions.
+        loss_v2t = F.cross_entropy(logits, targets)
+        loss_t2v = F.cross_entropy(logits.t(), targets)
+        loss = 0.5 * (loss_v2t + loss_t2v)
+        return loss
+
+###############################################################################
+# DDP-aware version: uses a custom autograd gather to preserve gradients.
+###############################################################################
+
+class GatherLayer(torch.autograd.Function):
+    """
+    Custom autograd function that performs an all_gather on the input tensor
+    while preserving gradients.
+    """
+    @staticmethod
+    def forward(ctx, tensor: torch.Tensor) -> torch.Tensor:
+        if not dist.is_initialized():
+            return tensor
+        world_size = dist.get_world_size()
+        output = [torch.zeros_like(tensor) for _ in range(world_size)]
+        # Synchronously gather tensors from all processes.
+        dist.all_gather(output, tensor)
+        return torch.cat(output, dim=0)
+    
+    @staticmethod
+    def backward(ctx, grad_output: torch.Tensor) -> torch.Tensor:
+        if not dist.is_initialized():
+            return grad_output
+        world_size = dist.get_world_size()
+        local_rank = dist.get_rank()
+        # Split the gradient and return the chunk corresponding to this process.
+        grad_list = grad_output.chunk(world_size, dim=0)
+        return grad_list[local_rank]
+
+def gather_all(tensor: torch.Tensor) -> torch.Tensor:
+    """
+    Gathers a tensor from all DDP processes using the custom GatherLayer.
+    Assumes tensor shape is [local_batch_size, ...].
+    """
+    if not dist.is_initialized():
+        return tensor
+    return GatherLayer.apply(tensor)
+
+@LossRegistry.register('contrastive_ddp')
+class ContrastiveLossDDP(nn.Module):
+    """
+    CLIP-style bidirectional contrastive loss for video and text embeddings in DDP setting.
+    """
+    def __init__(self):
+        """
+        Initialize DDP contrastive loss.
+        """
+        super().__init__()
+    
+    def forward(
+        self,
+        video_features: torch.Tensor,
+        text_features: torch.Tensor,
+        log_temp: torch.Tensor = torch.log(torch.tensor(0.1))
+    ) -> torch.Tensor:
+        """
+        Compute contrastive loss between video and text features in DDP setting.
+        
+        Args:
+            video_features (torch.Tensor): [B, D] video embeddings (local batch)
+            text_features (torch.Tensor): [B, D] text embeddings (local batch)
+            log_temp (torch.Tensor, optional): Override internal temperature
+            
+        Returns:
+            torch.Tensor: Scalar loss value
+        """        
+        # 1) Gather features from all GPUs.
+        video_features_all = gather_all(video_features)
+        text_features_all  = gather_all(text_features)
+
+        # 2) Normalize the gathered features.
+        video_features_all = F.normalize(video_features_all, dim=1)
+        text_features_all  = F.normalize(text_features_all, dim=1)
+
+        # 3) Compute global similarity matrix of shape [N, N],
+        # where N is the total global batch size.
+        similarity_matrix = torch.matmul(video_features_all, text_features_all.t())
+
+        # 4) Apply temperature scaling.
+        temp = torch.exp(log_temp)
+        logits = similarity_matrix / temp
+
+        # 5) Create targets: assume that the matching pairs lie along the diagonal.
+        n = logits.size(0)  # global batch size
+        targets = torch.arange(n, device=logits.device)
+
+        # Compute cross-entropy losses in both directions.
+        loss_i2t = F.cross_entropy(logits, targets)
+        loss_t2i = F.cross_entropy(logits.t(), targets)
+        loss = 0.5 * (loss_i2t + loss_t2i)
+        return loss
+
+@LossRegistry.register('InfoNCE')
+class InfoNCELoss(nn.Module):
+    """
+    InfoNCE loss implementation for contrastive learning.
+    This is a wrapper around the contrastive loss functions that maintains state
+    and provides a more object-oriented interface.
+    """
+    def __init__(self, temperature: float = 0.07, use_ddp: bool = False):
+        """
+        Initialize InfoNCE loss.
+        
+        Args:
+            temperature (float): Temperature parameter for scaling logits
+            use_ddp (bool): Whether to use DDP-aware version of the loss
+        """
+        super().__init__()
+        self.temperature = temperature
+        self.use_ddp = use_ddp
+        self.log_temp = nn.Parameter(torch.log(torch.tensor(temperature)))
+        
+    def forward(
+        self,
+        video_features: torch.Tensor,
+        text_features: torch.Tensor,
+        log_temp: torch.Tensor = None
+    ) -> torch.Tensor:
+        """
+        Compute InfoNCE loss between video and text features.
+        
+        Args:
+            video_features (torch.Tensor): [B, D] video embeddings
+            text_features (torch.Tensor): [B, D] text embeddings
+            log_temp (torch.Tensor, optional): Override internal temperature
+            
+        Returns:
+            torch.Tensor: Scalar loss value
+        """
+        # Use provided temperature if given, otherwise use internal one
+        temp = log_temp if log_temp is not None else self.log_temp
+        
+        if self.use_ddp and dist.is_initialized():
+            return ContrastiveLossDDP()(video_features, text_features, temp)
+        else:
+            return ContrastiveLoss()(video_features, text_features, temp)
+
